Use ptrdiff_t on Solaris and illumos. [skip appveyor]
In this particular context the two pointers point within the same
interface name string hence the difference always fits into an int, but
Clang needs it done right.  The problem reproduced on illumos too.

(cherry picked from commit 2387e6f3f58172e69e8d3fa087352275da892f43)

+++ pcap-bpf.c
@@ -120,6 +120,7 @@ static int bpf_load(char *errbuf);
 #include <stdlib.h>
 #include <string.h>
 #include <unistd.h>
+#include <stddef.h>
 
 #ifdef SIOCGIFMEDIA
 # include <net/if_media.h>
@@ -655,7 +656,7 @@ bpf_bind(int fd, const char *name, char *errbuf)
 	 */
 	if ((zonesep = strchr(name, '/')) != NULL) {
 		char *zname;
-		int  znamelen;
+		ptrdiff_t znamelen;
 
 		if (ifr.lifr_zoneid != GLOBAL_ZONEID) {
 			/*
